requirements.txt
asgiref==3.11.0
dj-rest-auth==7.0.1
Django==5.2
djangorestframework==3.16.1
et_xmlfile==2.0.0
openpyxl==3.1.5
pillow==12.0.0
qrcode==7.3
sqlparse==0.5.4
typing_extensions==4.15.0